// front/src/routes/routes.enum.ts
export enum AppRoutes {
  LOGIN = "/login",
  HOME = "/home",
  COMPANY_SETUP = "/company-setup",
  COMPANY_PICKER = "/company-picker",
  CALENDAR = "/calendar",
  CALENDAR2 = "/calendar2",
  RECIPES = "/recipes",
  TAGS = "/tags",
  EMPLOYEES = "/employees",
  PENDING_PREPARATIONS = "/pending-preparations",
  ACTIVITY_LOG = "/activity-log",
}